[FEAT/#88] service 파일 분리
