+++ src/main/java/com/example/picknwhip_be/domain/favorite/service/DesignMyPickCheckerService.java
@@ -3,6 +3,6 @@ package com.example.picknwhip_be.domain.favorite.service;
 import java.util.List;
 import java.util.Set;
 
-public interface DesignMyPickChecker {
+public interface DesignMyPickCheckerService {
   Set<Long> findPickedDesignIds(Long userId, List<Long> designIds);
 }
