+++ src/main/java/com/example/picknwhip_be/domain/favorite/service/DesignMyPickCheckerServiceImpl.java
@@ -3,13 +3,11 @@ package com.example.picknwhip_be.domain.favorite.service;
 import java.util.Collections;
 import java.util.List;
 import java.util.Set;
-import org.springframework.context.annotation.Primary;
+
 import org.springframework.stereotype.Component;
 
-// TODO: 추후에 마이픽 케이크 디자인 구현 후 @Primary 삭제
 @Component
-@Primary
-public class NoOpDesignMyPickChecker implements DesignMyPickChecker {
+public class DesignMyPickCheckerServiceImpl implements DesignMyPickCheckerService {
 
   @Override
   public Set<Long> findPickedDesignIds(Long userId, List<Long> designIds) {
